Switched to simpler functional api. With export and import feature.
There is no central model class with state anymore.
A model is defined by calling a create model and loss and training it with a train function.

This simplifies merging multiple models and avoids wrong usage of the api.

Also an experimental export and import function for graphs is implemented.
This should simplify deployment.